requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-environ==0.10.0
django-extensions==3.2.1
django-htmx==1.14.0
idna==3.4
Pillow==9.5.0
requests==2.28.2
sqlparse==0.4.3
stripe==5.4.0
urllib3==1.26.15
